Fix bug been locked near lower zoom limit on windows.
Trying to zoom in with mouse wheel after reaching the lower limit
locks the zoom level at 11%.

For better debugging support print out current zoom level on any update.

CCBUG:358097

+++ umbrello/umlview.h
@@ -40,8 +40,8 @@ public:
 
     UMLScene* umlScene() const;
 
-    int zoom() const ;
-    void setZoom(int zoom);
+    qreal zoom() const ;
+    void setZoom(qreal zoom);
 
     virtual bool showPropertiesDialog(QWidget *parent = 0);
